Revert "스킬오브젝트 비활성화 조건 수정"
This reverts commit 70f0340fa608b28b326037a2df0b0c16a363fc59.

+++ LevelManager.cpp
@@ -143,7 +143,7 @@ void LevelManager::DeleteObject(GameObject& me)
 	assert(curLevel);
 
 	auto& objects = curLevel->objects[(int)me.GetLayer()];
-	size_t size = objects.size();
+	int size = objects.size();
 
 	for (int i=0; i<size; ++i)
 	{
